feature flag rework
