+++ profiling-procmacros/Cargo.toml
@@ -19,3 +19,10 @@ syn = { version = "1.0", features = ["full"] }
 
 [lib]
 proc-macro=true
+
+[features]
+profile-with-puffin = []
+profile-with-optick = []
+profile-with-superluminal = []
+profile-with-tracing = []
+profile-with-tracy = []
