+++ profiling-procmacros/src/lib.rs
@@ -12,29 +12,7 @@ pub fn function(
     let instrumented_function_name = function.sig.ident.to_string();
 
     let body = &function.block;
-    let new_body: syn::Block = parse_quote! {
-        {
-            #[cfg(feature = "profile-with-puffin")]
-            profiling::puffin::profile_function!();
-
-            #[cfg(feature = "profile-with-optick")]
-            profiling::optick::event!();
-
-            #[cfg(feature = "profile-with-tracing")]
-            let _fn_span = profiling::tracing::span!(tracing::Level::INFO, #instrumented_function_name);
-            #[cfg(feature = "profile-with-tracing")]
-            let _fn_span_entered = _fn_span.enter();
-
-            #[cfg(feature = "profile-with-tracy")]
-            // Note: callstack_depth is 0 since this has significant overhead
-            let _tracy_span = profiling::tracy_client::Span::new(#instrumented_function_name, "", file!(), line!(), 0);
-
-            #[cfg(feature = "profile-with-superluminal")]
-            let _superluminal_guard = profiling::superluminal::SuperluminalGuard::new(#instrumented_function_name);
-
-            #body
-        }
-    };
+    let new_body: syn::Block = impl_block(body, &instrumented_function_name);
 
     function.block = Box::new(new_body);
 
@@ -42,4 +20,94 @@ pub fn function(
         #function
     })
     .into()
+}
+
+#[cfg(not(any(
+    feature = "profile-with-puffin",
+    feature = "profile-with-optick",
+    feature = "profile-with-superluminal",
+    feature = "profile-with-tracing",
+    feature = "profile-with-tracy",
+)))]
+fn impl_block(
+    body: &syn::Block,
+    _instrumented_function_name: &str,
+) -> syn::Block {
+    parse_quote! {
+        {
+            #body
+        }
+    }
+}
+
+#[cfg(feature = "profile-with-puffin")]
+fn impl_block(
+    body: &syn::Block,
+    _instrumented_function_name: &str,
+) -> syn::Block {
+    parse_quote! {
+        {
+            profiling::puffin::profile_function!();
+
+            #body
+        }
+    }
+}
+
+#[cfg(feature = "profile-with-optick")]
+fn impl_block(
+    body: &syn::Block,
+    _instrumented_function_name: &str,
+) -> syn::Block {
+    parse_quote! {
+        {
+            profiling::optick::event!();
+
+            #body
+        }
+    }
+}
+
+#[cfg(feature = "profile-with-superluminal")]
+fn impl_block(
+    body: &syn::Block,
+    instrumented_function_name: &str,
+) -> syn::Block {
+    parse_quote! {
+        {
+            let _superluminal_guard = profiling::superluminal::SuperluminalGuard::new(#instrumented_function_name);
+
+            #body
+        }
+    }
+}
+
+#[cfg(feature = "profile-with-tracing")]
+fn impl_block(
+    body: &syn::Block,
+    instrumented_function_name: &str,
+) -> syn::Block {
+    parse_quote! {
+        {
+            let _fn_span = profiling::tracing::span!(tracing::Level::INFO, #instrumented_function_name);
+            let _fn_span_entered = _fn_span.enter();
+
+            #body
+        }
+    }
+}
+
+#[cfg(feature = "profile-with-tracy")]
+fn impl_block(
+    body: &syn::Block,
+    instrumented_function_name: &str,
+) -> syn::Block {
+    parse_quote! {
+        {
+            // Note: callstack_depth is 0 since this has significant overhead
+            let _tracy_span = profiling::tracy_client::Span::new(#instrumented_function_name, "", file!(), line!(), 0);
+
+            #body
+        }
+    }
 }
